requirements.txt
ffmpeg-python==0.2.0
git+https://github.com/JavDomGom/pytube
